feat(run): add initial spike of a tracer
there is lots here that can be cleaned up but we
get a trace_id and span_id generated and
passed to the command. I have validated that
this can forward to an opentelemetry-collector
(otelcol-contrib v0.44.0) and forward the logs
into datadog.  there are a few problems still
where tags are not translating properly in all
cases and inner spans are not linked. I'm going
to work through which parts of this are datadog
and which parts are opentelemetry