fix: improve predictability of command parsing
the original method of parsing command args was overly simplistic
assuming that the command would be sent as a single string and
opentracer would have to parse it, splitting on spaces

this wound up with some unpredictable interaction between shell
interpolation and quoted args with spaces in them

to simplify this we are now leaving command-line parsing to the
cobra library and the shell.

this means that opentracer no longer requires the shell command
to be a single string (which necessitated extra wrangling to properly
preserve quotes and spaces across shell interpretation) but now
supports any number of args and flags passed to that inner shell
command without

README and long-form help text in the run command have been
updated to better explain how this opentracer parses and passes
commands.

hopefully the new conventions will be more familiar and predictable
to experienced shell users

BREAKING CHANGE: existing run args should be reviewed and updated for compatibility with improved arg parsing