fix(build): release task failed with dirty changes
with the test task as a depenency on the relase task
the test task (with its dependent version file updates)
was getting invoked first so that the release failed
seeing the updated version files as dirty.

this commit moves the test task to after the release
task's preconditions allowing the release precondition
checks to succeed on an empty repo before the version
files are updated.